DDF-2773: Remove SourceInfoField
