+++ query/sources/impl/src/main/java/org/codice/ddf/admin/sources/opensearch/discover/GetOpenSearchConfigurations.java
@@ -97,7 +97,7 @@ public class GetOpenSearchConfigurations
         });
 
         for (OpenSearchSourceInfoField sourceInfoField : cswSourceInfoFields.getList()) {
-            sourceUtilCommons.populateAvailability(sourceInfoField,
+            sourceUtilCommons.populateAvailability(sourceInfoField.isAvailableField(),
                     sourceInfoField.config()
                             .pidField());
             sourceInfoField.config()
